requirements.txt
streamlit 
pdfminer.six 
pdf2image 
openai 
pillow
python-dotenv
pdfplumber
requests
img2pdf
PyMuPDF
pymupdf4llm
markdown
bs4
camelot-py